detached python backend entirely to another process.

+++ tracer/trace_file.h
@@ -6,6 +6,6 @@
 void trace_file_init(void);
 void trace_file_begin(void);
 void trace_file_end(void);
-void trace_file_log(target_ulong vaddr, target_ulong paddr, uint32_t flags, uint64_t icount);
+void trace_file_log(target_ulong vaddr, target_ulong paddr, uint64_t flags, uint64_t icount);
 
 #endif // __STATISTIC_FILE_H__
